modificado:    Makefile 		Adicionados alguns ficheiros.
	novo ficheiro: include/Engine.h
	novo ficheiro: games/Engine.c
		Ficheiro responsavel por ser o mediador entre jogos e jogador.

	modificado:    include/guess_game.h
	modificado:    games/guess_game/guess_game.c
		Modificada a logica de input.

	modificado:    games/jogo_do_galo/jogo_do_galo.c
		Tornou-s eo texto de menu constante.

	modificado:    include/aux_func.h
	modificado:    utils/aux_func.c
		Adicionadas novas funções entre elas a nova função global de input de inteiros.

	novo ficheiro: include/jogador.h
	modificado:    include/main_menu.h
	modificado:    include/setup.h
	modificado:    menu/main_menu.c
		Mudada a lógica de input.

	modificado:    src/main.c
	modificado:    src/setup.c
		Revertido para int setup.

+++ include/main_menu.h
@@ -4,5 +4,5 @@
  * @brief Menu inicial.
  * @return 0 se correu tudo bem, 1 se ouve um problema.
  */
-signed short int mainMenu();
+signed int mainMenu();
 #endif
